posting_api/tests/functional/requirements.txt
psycopg2-binary==2.8.6
pytest==6.1.2
python-decouple==3.3
sqlalchemy==1.3.20
